Catch up with Scala Library Modularization.
Split the dependencies along the scala-2.10.x and scala-2.11.x profiles.
The goal is to evolve the 2.11 module structure to get a leaner set of
dependencies on Scala Core, with a separate Scala Modules container.

Support the 2.11 swing/continuations modules:
Rename `continuations.jar` to `scala-continuations-plugin.jar`.
In 2.10 this was called `continuations.jar`,
in 2.11 it's `scala-continuations-plugin_2.11.jar`, and the library part
is distributed separately, so add that to the Scala container.

TODO: add scala-xml and scala-parser-combinators to the scala container and to the repl classpath

(These dependencies should move to the Scala Modules container eventually.)

Fix the outputDirectory-es of scala-(refactoring|swing)